refactor: tidy up <SwitchAccount /> component
